requirements.txt

#docs
sphinx
recommonmark
commonmark
sphinx-autobuild
repoze.sphinx.autointerface
doc8
autodoc

#devtool
autopep8

# component
zope.component

# gui
pyqt5
qtpy
rx

# for jupyter notebook
asyncqt

# console
qtconsole

# traits
traits
traitsui

matplotlib

# datetime
pytz
iso8601
datetime

# fp
first


#graph
pyqtgraph

# image
scikit-image
scipy 